Move getHttpStatus to own file to satisfy bundle size limit
