+++ src/auth/utils.ts
@@ -4,16 +4,6 @@ import { API_URL } from "./constants";
 export const sleep = (ms: number) =>
   new Promise((resolve) => setTimeout(resolve, ms));
 
-/**
- * Extract the HTTP status code from an SDK API error message.
- * Returns `undefined` if the error does not match the `"API error (NNN)"` pattern.
- */
-export const getHttpStatus = (error: unknown): number | undefined => {
-  if (!(error instanceof Error)) return undefined;
-  const match = error.message.match(/API error \((\d+)\)/);
-  return match ? parseInt(match[1], 10) : undefined;
-};
-
 export async function authRequest<T>(
   endpoint: string,
   options: RequestInit = {},
